Ulong Comparison
| Method             | Mean     | Error    | StdDev   | Allocated |
|------------------- |---------:|---------:|---------:|----------:|
| NewNaturalComparer | 469.1 us |  4.01 us |  3.55 us |         - |
| OldNaturalComparer | 930.3 us | 11.38 us | 10.08 us |         - |

BigInteger Comparison
| Method             | Mean       | Error     | StdDev    | Gen0      | Allocated |
|------------------- |-----------:|----------:|----------:|----------:|----------:|
| NewNaturalComparer |   700.0 us |   4.95 us |   3.86 us |         - |         - |
| OldNaturalComparer | 6,610.9 us | 128.89 us | 167.59 us | 1023.4375 | 4287331 B |